backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Libraries
sentence-transformers==2.2.2
spacy==3.7.2
openai==1.3.7
transformers==4.35.2
torch==2.1.1

# Vector Database
chromadb==0.4.18

# Graph Processing
networkx==3.2.1

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Utilities
python-dateutil==2.8.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2

# WebSocket
websockets==12.0

# CORS
python-cors==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
